Revert "zpj-miner: don't descend into folders that didn't change"
Unfortunately this doesn't really work - for the way we clean up deleted
resources we need to have a complete list. We could query the missing
pieces from the database, but the fact that we would have to recurse
with a query roundtrip each time makes me wonder if it's worth it.
For now, just revert the commit.

This reverts commit 68c88e8f5627835644fd916d3966d9842a2fd7a4.