application: don't use Signals.addSignalMethods
It doesn't work properly when used together with a GObject subclass that
emits its own signals.
As a workaround, add a addJSSignalMethods() utility that adds methods
with different names explicitly for JS signals, and use it for JS
signals installed on the Application prototype.

This fixes application actions not being activated anymore after the
last commit.