searchbar: ignore action-state-changed signals which are not for us
I'm pretty sure this is a regression lower in the stack, probably GLib
or GJS; we connect to the detailed 'action-state-changed::search' signal
on the GApplication action group, but our callback is also invoked when
other action's state changes.
Workaround this by ignoring all the emissions that are not for us.