application: don't immediately forward search terms on item launch
When an item is launched from a shell search, we also set the search
string to the terms that were used.
Instead of doing so immediately, wait until the next overview state
change, since setString() will clear DocumentManager, and we have no
guarantees on the order clear and load will happen.

https://bugzilla.gnome.org/show_bug.cgi?id=697034