miner: add a versioning system to the mining operation
It's useful to have a way to associate a full reindex of an online
resource to a programmatic change in the code, since we avoid
updating information for entries that haven't changed by looking at
their mtime.

Add support for this. Right now the version of all the miners is set to
one, and is one by default, so nothing will be reindexed just yet.

https://bugzilla.gnome.org/show_bug.cgi?id=696928